# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talim AI - Backend для тестирования компетенций",
    "description": "\n    API для системы AI-тестирования профессиональных компетенций через голосовое собеседование.\n    \n    ## Основные возможности\n    \n    * **Тестирование компетенций**: Адаптивное голосовое тестирование с использованием GPT-4\n    * **Транскрипция аудио**: Автоматическая транскрипция ответов через Whisper API\n    * **Оценка ответов**: AI-оценка с выявлением пробелов в знаниях\n    \n    ## Документация\n    \n    * Swagger UI: `/docs`\n    * ReDoc: `/redoc`\n    * OpenAPI JSON: `/openapi.json`\n    \n    ## Аутентификация\n    \n    User ID передается с фронтенда в заголовке `Authorization`:\n    - Формат: `Authorization: Bearer {user_id}` или `Authorization: {user_id}`\n    - Заголовок обязателен для всех защищенных endpoints\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Получить все роли",
        "description": "Возвращает список всех доступных ролей в системе",
        "operationId": "get_roles_api_roles_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assessments/directions/{direction_id}/technologies": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Получить технологии направления",
        "description": "Возвращает список технологий для указанного направления",
        "operationId": "get_direction_technologies_api_assessments_directions__direction_id__technologies_get",
        "parameters": [
          {
            "name": "direction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Direction Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "default": "ID направления",
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/directions/{direction_id}/competencies": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Получить компетенции направления",
        "description": "Возвращает список общих компетенций для указанного направления",
        "operationId": "get_direction_competencies_api_assessments_directions__direction_id__competencies_get",
        "parameters": [
          {
            "name": "direction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Direction Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "default": "ID направления",
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/technologies/{technology_id}/competencies": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Получить компетенции технологии",
        "description": "Возвращает список компетенций для указанной технологии",
        "operationId": "get_technology_competencies_api_assessments_technologies__technology_id__competencies_get",
        "parameters": [
          {
            "name": "technology_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technology Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "default": "ID технологии",
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Начать новое тестирование",
        "description": "Создает новое тестирование для указанного направления и возвращает список компетенций для тестирования.",
        "operationId": "create_assessment_api_assessments_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Получить список тестирований пользователя",
        "description": "Возвращает список всех тестирований пользователя, отсортированных по номеру попытки и дате",
        "operationId": "get_user_assessments_api_assessments_get",
        "parameters": [
          {
            "name": "direction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по направлению",
              "title": "Direction Id"
            },
            "description": "Фильтр по направлению"
          },
          {
            "name": "technology_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по технологии",
              "title": "Technology Id"
            },
            "description": "Фильтр по технологии"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentResponse"
                  },
                  "title": "Response Get User Assessments Api Assessments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{assessment_id}": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Получить информацию о тестировании",
        "description": "Возвращает детальную информацию о тестировании, включая статус, оценки по компетенциям и прогресс",
        "operationId": "get_assessment_api_assessments__assessment_id__get",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assessment Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "default": "ID тестирования",
              "title": "Description"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{assessment_id}/complete": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Завершить тестирование",
        "description": "Завершает тестирование, вычисляет общий балл и устанавливает статус 'completed'",
        "operationId": "complete_assessment_api_assessments__assessment_id__complete_post",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assessment Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "default": "ID тестирования",
              "title": "Description"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentCreate": {
        "properties": {
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "technology": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology"
          }
        },
        "type": "object",
        "required": [
          "direction"
        ],
        "title": "AssessmentCreate"
      },
      "AssessmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "role_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "direction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction Id"
          },
          "technology_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number",
            "default": 1
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "competency_assessments": {
            "items": {
              "$ref": "#/components/schemas/CompetencyAssessmentResponse"
            },
            "type": "array",
            "title": "Competency Assessments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "status",
          "started_at"
        ],
        "title": "AssessmentResponse"
      },
      "AssessmentStartResponse": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assessment Id"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyInfo"
            },
            "type": "array",
            "title": "Competencies"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "competencies",
          "status"
        ],
        "title": "AssessmentStartResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}